+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Version 0.3
+
+ * Hover overs seem to work!
+ * Added UnityNova to project for using Nova with Unity.
+ * Activates when opening a *.cs file.
+ * Started messing around with preferences.
+
 ## Version 0.2
 
 Should now have Symbols in Nova!